게시글 삭제 성공 - AJAX요청에 대한 공통DTO생성, 공통Exception파일 생성

+++ src/main/java/com/cos/blogapp/web/BoardController.java
@@ -17,107 +17,120 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.cos.blogapp.domain.board.Board;
 import com.cos.blogapp.domain.board.BoardRepository;
 import com.cos.blogapp.domain.user.User;
+import com.cos.blogapp.handler.exception.MyAsyncNotFoundException;
 import com.cos.blogapp.handler.exception.MyNotFoundException;
 import com.cos.blogapp.util.Script;
 import com.cos.blogapp.web.dto.BoardSaveReqDto;
+import com.cos.blogapp.web.dto.CMRespDto;
 
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor // final이 붙은 필드에 대한 생성자가 만들어진다.
 @Controller // 컴포넌트 스캔(스프링) IoC ()
 public class BoardController {
-	
+
 	// DI
 	private final BoardRepository boardRepository;
 	private final HttpSession session;
-	
-	// RestFul API 주소 설계 방식
-//	@GetMapping("/board/1/comment")
-	
-	// DELETE FROM board WHERE id = ?
+
+	// API(AJAX) 요청
 	@DeleteMapping("/board/{id}")
-	public @ResponseBody String deleteByID(@PathVariable int id) {
-		boardRepository.deleteById(id);
-		return "ok"; // @ResponseBody 데이터 리턴!! String = text/plain
+	public @ResponseBody CMRespDto<String> deleteByID(@PathVariable int id) {
+
+		// 인증이 된 사람만 함수 접근 가능!! (로그인 된 사람)
+		User principal = (User) session.getAttribute("principal");
+
+		if (principal == null) { // 로그인 안됨
+			throw new MyAsyncNotFoundException("인증이 되지 않았습니다.");
+		}
+
+		// 권한이 있는 사람만 함수 접근 가능(principal.id == {id})
+		Board boardEntity = boardRepository.findById(id)
+				.orElseThrow(() -> new MyAsyncNotFoundException("해당 글을 찾을 수가 없습니다."));
+		if (principal.getId() != boardEntity.getUser().getId()) {
+			throw new MyAsyncNotFoundException("해당글을 삭제할 권한이 없습니다.");
+		}
+
+		try {
+			boardRepository.deleteById(id);
+		} catch (Exception e) {
+			throw new MyAsyncNotFoundException(id + "를 찾을 수 없어서 삭제할 수 없어요.");
+		}
+		return new CMRespDto<String>(1, "성공", null); // CMRespDto 는 object를 리턴 => JSON 형태
 	}
-	
+
 	// UPDATE board SET title = ?, content = ? WHERE id = ?
 	// @PutMapping("/board/{id}")
-	
+
 	// 쿼리스트링, 패스var => where 에 걸리는 친구들
 	// 1. 컨트롤러 선정, 2. Http Method 선정, 3. 받을 데이터가 있는지!! ( body, 쿼리스트링, 패스var )
 	// 4. DB에 접근을 해야하면 Model 접근하기 orElse Model에 접근할 필요가 없다.
 	@GetMapping("/board/{id}")
 	public String detail(@PathVariable int id, Model model) {
 		// select * from board where id = :id
-		
 		// 1. orElse 는 값을 찾으면 Board가 리턴, 못찾으면 (괄호안 내용 리턴)
 //		Board boardEntity = boardRepository.findById(id)
 //				.orElse(new Board(100, "글없어요", "글없어요", null));
-		
+
 		// 2. orElseThrow
-		Board boardEntity = boardRepository.findById(id)
-				.orElseThrow(()-> new MyNotFoundException(id+" 못찾았어요") );
-				// {} 사용 시 여러 줄 작성 가능, but return 삽입
-				// {} 생략 시 한줄만 작성 가능, return 생략
-		
+		Board boardEntity = boardRepository.findById(id).orElseThrow(() -> new MyNotFoundException(id + " 못찾았어요"));
+		// {} 사용 시 여러 줄 작성 가능, but return 삽입
+		// {} 생략 시 한줄만 작성 가능, return 생략
+
 		model.addAttribute("boardEntity", boardEntity);
 		return "board/detail";
 	}
-	
+
 	@PostMapping("/board")
 	public @ResponseBody String save(@Valid BoardSaveReqDto dto, BindingResult bindingResult) {
-		
+
 		User principal = (User) session.getAttribute("principal");
-		
+
 		// 인증, 권한체크(공통로직)
-		if(principal == null) { // 로그인 안됨
+		if (principal == null) { // 로그인 안됨
 			return Script.href("/loginForm", "잘못된 접근입니다");
 		}
-		
-		if( bindingResult.hasErrors() ) {
+
+		if (bindingResult.hasErrors()) {
 			Map<String, String> errorMap = new HashMap<>();
-			for(FieldError error : bindingResult.getFieldErrors()) {
+			for (FieldError error : bindingResult.getFieldErrors()) {
 				errorMap.put(error.getField(), error.getDefaultMessage());
-			}	
+			}
 			return Script.back(errorMap.toString());
-		} 
+		}
 		System.out.println("title : " + dto.getTitle());
 //		dto.setContent(dto.getContent().replaceAll("<p>",""));
 //		dto.setContent(dto.getContent().replaceAll("</p>",""));
 		System.out.println("content : " + dto.getContent());
-		
-		
+
 //		User user = new User();
 //		user.setId(3);
 //		boardRepository.save(dto.toEntity(user));
-		
+
 		boardRepository.save(dto.toEntity(principal));
-		return Script.href("/","글쓰기 성공");
+		return Script.href("/", "글쓰기 성공");
 	}
-	
+
 	@GetMapping("/board/saveForm")
 	public String saveForm() {
 		return "board/saveForm";
 	}
-	
+
 	// /board?page=1
 	@GetMapping("/board")
 	public String home(Model model, Integer page) {
 		PageRequest pageRequest = PageRequest.of(page, 3, Sort.by(Sort.Direction.DESC, "id"));
-		
-		//Sort.by(Sort.Direction.DESC, "id")
-		Page<Board> boardsEntity = 
-				boardRepository.findAll(pageRequest);
+
+		// Sort.by(Sort.Direction.DESC, "id")
+		Page<Board> boardsEntity = boardRepository.findAll(pageRequest);
 		model.addAttribute("boardsEntity", boardsEntity);
 		// System.out.println(boardsEntity.get(0).getUser().getUsername());
 		return "board/list";
 	}
-	
+
 }
